# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mocker admin API",
    "version": "1.0.0",
    "description": "The admin plane of mocker: everything the web UI talks to.\n\nThis document is the CONTRACT, not documentation-after-the-fact. web/ generates its\nentire API client from it (orval), and internal/admin/openapi_contract_test.go fails\nthe Go build's test bar if a single (method, path) here disagrees with the routes\ninternal/admin actually registers. Adding a handler without adding it here — or the\nreverse — is a red test, not a silent drift.\n\nEvery route below is served on the admin host (MOCKER_ADMIN_HOST) only. The mock\nplane (per-workspace hosts) is a different surface entirely and is NOT described here."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the UI: the SPA is served by this very binary, so the client never needs an absolute base URL."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mocker_session",
        "description": "Set by POST /api/auth/login. HttpOnly, SameSite=Lax, and Secure unless MOCKER_DEV=1."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required on every POST/PUT/PATCH/DELETE except login. Value comes from the login/me response's csrfToken. Compared in constant time. Two further checks apply to the same requests and are NOT expressible here: Content-Type must parse to exactly application/json, and Origin (or Referer, when Origin is absent) must have hostname == MOCKER_ADMIN_HOST."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Malformed request: unparseable body, an out-of-range field, an invalid path parameter.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "CSRF refusal: a missing/incorrect X-CSRF-Token, or an Origin/Referer whose hostname is not MOCKER_ADMIN_HOST.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "UnsupportedMediaType": {
        "description": "Content-Type did not parse to application/json.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Internal": {
        "description": "Server-side failure. The message is deliberately generic; the detail is in the server log.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The single error envelope every non-2xx answer on this plane uses (internal/httpx/respond.go).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine code. The shared set is bad_request, unauthorized, forbidden, not_found, conflict, too_large, internal; session's GET/POST/DELETE add service_unavailable (no LiveStateSource wired in this deployment), and login adds rate_limited and unsupported_media_type. not_implemented_yet does not appear anywhere in this contract: it named DESIGN §12's scenario switch before scenarios existed (POST .../session {\"scenario\":...} answered 501), and stopped being true the moment scenarios shipped their own activate/deactivate routes (A17) — that key is a plain 400 here now. A3 adds edit_conflict on the five per-row compare-and-swap routes (PATCH /api/workspaces/{id}, PUT .../operations/{opKey}, POST .../auth-preset, PUT .../endpoints/{eid}, PUT .../scenarios/{sid}): the caller's editVersion expectation no longer matches the row's stored edit_version.",
                "examples": [
                  "bad_request",
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "conflict",
                  "too_large",
                  "internal",
                  "rate_limited",
                  "unsupported_media_type",
                  "service_unavailable",
                  "edit_conflict"
                ]
              },
              "message": {
                "type": "string",
                "description": "Human text. Shown to the operator as-is."
              },
              "details": {
                "description": "Per-field validation detail. Absent unless the route produces one."
              }
            }
          }
        }
      },
...
      "UserView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "Trimmed, non-empty, no control characters, at most 64 runes — enforced server-side. A UI that pre-validates must use exactly these rules and no others."
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "Unix seconds."
          }
        }
      },
...
      "ServerConfigView": {
        "type": "object",
        "description": "Server-side configuration the browser cannot derive on its own. MOCKER_ADMIN_HOST is forbidden from sitting under MOCKER_BASE_DOMAIN, so window.location tells the UI nothing about either the base domain or the reserved prefix.",
        "required": [
          "reservedPrefix",
          "baseDomain",
          "routing",
          "limits"
        ],
        "properties": {
          "reservedPrefix": {
            "type": "string",
            "description": "MOCKER_RESERVED_PREFIX — the path prefix a workspace's mock plane answers its own control routes on ({prefix}/health, {prefix}/state). Never hard-code /__mocker in the UI: it is configurable and the acceptance run uses a different value on purpose."
          },
          "baseDomain": {
            "type": "string",
            "description": "MOCKER_BASE_DOMAIN — the domain workspace hosts live under in host routing mode."
          },
          "routing": {
            "type": "string",
            "enum": [
              "host",
              "path"
            ],
            "description": "MOCKER_ROUTING."
          },
          "limits": {
            "type": "object",
            "description": "A9: the effective ceilings of this process, the same config.Limits an agent reads through get_server_config. Bytes are bytes, seconds are seconds.",
            "required": [
              "maxBodyBytes",
              "maxResponseBytes",
              "maxAssetBytes",
              "maxAssetsTotalBytes",
              "maxEntities",
              "trafficMaxBodyBytes",
              "trafficRetention",
              "checkpointRetention",
              "checkpointDebounceSec",
              "streamMaxConns",
              "streamMaxLifetimeSec",
              "streamMaxFrameBytes",
              "streamSendBudgetBytes",
              "streamPingSec",
              "streamFrameTimeoutSec",
              "streamTrafficFrames",
              "streamTrafficMaxFrames",
              "streamTrafficMaxBytes"
            ],
            "properties": {
              "maxBodyBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_MAX_BODY"
              },
              "maxResponseBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_MAX_RESPONSE — a body, and one stream frame"
              },
              "maxAssetBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_MAX_ASSET"
              },
              "maxAssetsTotalBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_MAX_ASSETS_TOTAL"
              },
              "maxEntities": {
                "type": "integer",
                "description": "MOCKER_MAX_ENTITIES"
              },
              "trafficMaxBodyBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_TRAFFIC_MAX_BODY"
              },
              "trafficRetention": {
                "type": "integer",
                "description": "MOCKER_TRAFFIC_RETENTION"
              },
              "checkpointRetention": {
                "type": "integer",
                "description": "MOCKER_CHECKPOINT_RETENTION"
              },
              "checkpointDebounceSec": {
                "type": "integer",
                "description": "MOCKER_CHECKPOINT_DEBOUNCE"
              },
              "streamMaxConns": {
                "type": "integer",
                "description": "MOCKER_STREAM_MAX_CONNS, per workspace"
              },
              "streamMaxLifetimeSec": {
                "type": "integer",
                "description": "MOCKER_STREAM_MAX_LIFETIME"
              },
              "streamMaxFrameBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_STREAM_MAX_FRAME, an inbound WebSocket frame"
              },
              "streamSendBudgetBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_STREAM_SEND_BUDGET"
              },
              "streamPingSec": {
                "type": "integer",
                "description": "MOCKER_STREAM_PING"
              },
              "streamFrameTimeoutSec": {
                "type": "integer",
                "description": "MOCKER_STREAM_FRAME_TIMEOUT"
              },
              "streamTrafficFrames": {
                "type": "string",
                "enum": [
                  "off",
                  "first",
                  "all"
                ],
                "description": "MOCKER_STREAM_TRAFFIC_FRAMES — off | first | all (A14: every frame each way into the connection's one row, under the two budgets below)"
              },
              "streamTrafficMaxFrames": {
                "type": "integer",
                "description": "MOCKER_STREAM_TRAFFIC_MAX_FRAMES — frames per row each way under \"all\""
              },
              "streamTrafficMaxBytes": {
                "type": "integer",
                "format": "int64",
                "description": "MOCKER_STREAM_TRAFFIC_MAX_BYTES — bytes per row each way under \"all\""
              }
            }
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "description": "The shared answer of POST /api/auth/login and GET /api/me.",
        "required": [
          "user",
          "csrfToken",
          "config"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "csrfToken": {
            "type": "string",
            "description": "Send back as X-CSRF-Token on every state-changing call."
          },
          "config": {
            "$ref": "#/components/schemas/ServerConfigView"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "name",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "summary": "Open a session",
        "description": "The only state-changing route exempt from the X-CSRF-Token check — it has no session to take a token from yet. The Content-Type and Origin checks still apply.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session opened. The cookie is set on this response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "Invalid credentials. Deliberately the SAME answer for a wrong password and an unknown name — the UI must not distinguish them either.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "description": "Rate limited: 10 attempts per minute, bucketed by (client address, submitted name).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  }
}